docs/code/python_dataviz/streamlit/cog_viewer/requirements.txt
google-api-python-client
streamlit
leafmap
rioxarray
google-cloud-storage